Move persist docs tests to the adapter zone. Prep for adapter tests.
